generate exported-array in frontend; break up big block
