+++ edg_core/__init__.py
@@ -3,7 +3,7 @@ from .ConstraintExpr import BoolLike, FloatLike, IntLike, RangeLike, StringLike,
 from .ConstraintExpr import Default
 from .Ports import Port, Bundle
 from .NotConnectablePort import NotConnectableBlock, NotConnectablePort
-from .Blocks import Link
+from .Link import Link
 from .DesignTop import DesignTop
 from .HierarchyBlock import Block, ImplicitConnect, init_in_parent, abstract_block
 from .Generator import GeneratorBlock
